html/browsers/history/the-location-interface/location_protocol.html

<!DOCTYPE HTML>
<html>
  <head>
    <title>location_protocol</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <script>
    if (('location' in window) && window.location != null
      && ('location' in document) && document.location != null){
      var protocol = location.protocol;
      var url = location.href;

      var pos = url.indexOf("//");
      if (pos != -1) {
        url = url.substr(0, pos);
      }

      test(function() {
        assert_equals(protocol, url, "protocol attribute is not correct.");
      });
    } else {
      test(function() {
        assert_unreached("location interface is not supported.");
      });
    }
    </script>
  </body>
</html>
